Update unit tests to start by loading the kernel package.
This is an automated change that adds exec() to the start of each test.
Scilab switches into the root directory of a module when running unit
tests, so we require that the the bindings be built in a build/
subdirectory of the root directory, for lack of a way to allow arbitrary
build directories.

(There are a number of test failures now, 20% failing.)

Fixes https://github.com/khumba/cgal-swig-bindings/issues/3.